Add Services cloud firestore

+++ lib/screens/note_list_screen.dart
@@ -1,5 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
+import 'package:notes/services/note_service.dart';
 
 class NoteListScreen extends StatefulWidget {
   const NoteListScreen({super.key});
@@ -63,20 +64,26 @@ class _NoteListScreenState extends State<NoteListScreen> {
                   ),
                   ElevatedButton(
                       onPressed: () {
-                        Map<String, dynamic> newNote = {};
+                        // Map<String, dynamic> newNote = {};
+
                         // Ini adalah alternatif dari yang diatas
                         // Map<String, dynamic> newNote = new Map<String, dynamic>();
-                        newNote['title'] = _titleController.text;
-                        newNote['description'] = _descriptionController.text;
 
-                        FirebaseFirestore.instance
-                            .collection('notes')
-                            .add(newNote)
-                            .whenComplete(
-                          () {
-                            Navigator.of(context).pop();
-                          },
-                        );
+                        // newNote['title'] = _titleController.text;
+                        // newNote['description'] = _descriptionController.text;
+
+                        NoteService.addNote(_titleController.text,
+                                _descriptionController.text)
+                            .whenComplete(() => Navigator.of(context).pop());
+
+                        // FirebaseFirestore.instance
+                        //     .collection('notes')
+                        //     .add(newNote)
+                        //     .whenComplete(
+                        //   () {
+                        //     Navigator.of(context).pop();
+                        //   },
+                        // );
                         _titleController.clear();
                         _descriptionController.clear();
                       },
@@ -94,7 +101,6 @@ class _NoteListScreenState extends State<NoteListScreen> {
 }
 
 class NoteList extends StatelessWidget {
-
   const NoteList({super.key});
 
   @override
@@ -103,7 +109,7 @@ class NoteList extends StatelessWidget {
     TextEditingController descriptionController = TextEditingController();
 
     return StreamBuilder(
-        stream: FirebaseFirestore.instance.collection('notes').snapshots(),
+        stream: NoteService.getNoteList(),
         builder: (context, snapshot) {
           if (snapshot.hasError) {
             return new Text('Error: ${snapshot.error}');
@@ -116,19 +122,22 @@ class NoteList extends StatelessWidget {
             default:
               return ListView(
                   padding: const EdgeInsets.only(bottom: 80),
-                  children: snapshot.data!.docs.map((document) {
+                  children: snapshot.data!.map((document) {
                     return Card(
                       child: ListTile(
                         onTap: () {
                           showDialog(
                             context: context,
                             builder: (context) {
-                              TextEditingController titleController = TextEditingController(text: document['title']);
-                              TextEditingController descriptionController = TextEditingController(text: document['description']);
+                              TextEditingController titleController =
+                                  TextEditingController(
+                                      text: document['title']);
+                              TextEditingController descriptionController =
+                                  TextEditingController(
+                                      text: document['description']);
 
                               return AlertDialog(
                                 title: const Text('Update Notes'),
-
                                 content: Column(
                                   crossAxisAlignment: CrossAxisAlignment.start,
                                   children: [
@@ -137,7 +146,7 @@ class NoteList extends StatelessWidget {
                                       textAlign: TextAlign.start,
                                     ),
                                     TextField(
-                                      controller: titleController,    
+                                      controller: titleController,
                                     ),
                                     const Padding(
                                       padding: const EdgeInsets.only(top: 10),
@@ -166,19 +175,30 @@ class NoteList extends StatelessWidget {
                                   ),
                                   ElevatedButton(
                                       onPressed: () {
-                                        Map<String, dynamic> updateNote = {};
-                                        updateNote['title'] = titleController.text;
-                                        updateNote['description'] = descriptionController.text;
+                                        // Map<String, dynamic> updateNote = {};
+                                        // updateNote['title'] =
+                                        //     titleController.text;
+                                        // updateNote['description'] =
+                                        //     descriptionController.text;
 
-                                        FirebaseFirestore.instance
-                                            .collection('notes')
-                                            .doc(document.id) //--> sintax agar update berdasarkan id, kalo tidak ada ini maka akan mengupdate semua
-                                            .update(updateNote)
-                                            .whenComplete(
-                                          () {
-                                            Navigator.of(context).pop();
-                                          },
-                                        );
+                                        // FirebaseFirestore.instance
+                                        //     .collection('notes')
+                                        //     .doc(document
+                                        //         .id) //--> sintax agar update berdasarkan id, kalo tidak ada ini maka akan mengupdate semua
+                                        //     .update(updateNote)
+                                        //     .whenComplete(
+                                        //   () {
+                                        //     Navigator.of(context).pop();
+                                        //   },
+                                        // );
+
+                                        NoteService.updateNote(
+                                                document['id'],
+                                                titleController.text,
+                                                descriptionController.text)
+                                            .whenComplete(() =>
+                                                Navigator.of(context).pop());
+
                                         titleController.clear();
                                         descriptionController.clear();
                                       },
@@ -193,44 +213,44 @@ class NoteList extends StatelessWidget {
                         trailing: InkWell(
                           onTap: () {
                             showDialog(
-                            context: context,
-                            builder: (context) {
-                               return AlertDialog(
-                              content: Column(
-                                children: [
-                                  Text('Delete Data')
-                                ],
-                              ),
-
-                            actions: [
-                                  Padding(
-                                    padding: const EdgeInsets.symmetric(
-                                        horizontal:
-                                            30), //--> Mengatur jarak antar tombol
-                                    child: ElevatedButton(
-                                      onPressed: () {
-                                        Navigator.of(context).pop(); //--> Tombol Menutup dialog atau cancel
-                                      },
-                                      child: Text('Cancel'),
-                                    ),
+                              context: context,
+                              builder: (context) {
+                                return AlertDialog(
+                                  content: Column(
+                                    children: [Text('Delete Data')],
                                   ),
-                                  ElevatedButton(
-                                    onPressed: () {
-                                      FirebaseFirestore.instance
-                                        .collection('notes')
-                                        .doc(document.id)
-                                        .delete()
-                                        .catchError((e) {
-                                      print(e);
-                                    });
-                                    Navigator.of(context).pop();
-                                  },
-                                  child: const Text('Update')),
-                                ],
-                              );
-                            },
-                          );
-                        },
+                                  actions: [
+                                    Padding(
+                                      padding: const EdgeInsets.symmetric(
+                                          horizontal:
+                                              30), //--> Mengatur jarak antar tombol
+                                      child: ElevatedButton(
+                                        onPressed: () {
+                                          Navigator.of(context)
+                                              .pop(); //--> Tombol Menutup dialog atau cancel
+                                        },
+                                        child: Text('Cancel'),
+                                      ),
+                                    ),
+                                    ElevatedButton(
+                                        onPressed: () {
+                                          FirebaseFirestore.instance
+                                              .collection('notes')
+                                              .doc(document['id'])
+                                              .delete()
+                                              .catchError((e) {
+                                            print(e);
+                                          });
+                                          Navigator.of(context).pop();
+
+                                          NoteService.deleteNote(document['id']);
+                                        },
+                                        child: const Text('Delete')),
+                                  ],
+                                );
+                              },
+                            );
+                          },
                           child: const Padding(
                             padding: EdgeInsets.symmetric(vertical: 10),
                             child: Icon(Icons.delete),
